better testing

+++ Kull.GenericBackend.IntegrationTest/TestWebapplicationFactory.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Collections.Generic;
 using System.Data.SqlClient;
+using Kull.Data;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -20,15 +21,36 @@ namespace Kull.GenericBackend.IntegrationTest
             {
                 if (System.IO.File.Exists(System.IO.Path.Combine(dataPath, "GenericBackendTest.mdf")))
                 {
-                    return;
-                    /* Currently, do not delete db
-                    using (SqlConnection connection = new SqlConnection(@"server=(localdb)\MSSQLLocalDB"))
+                    string testCommand = "SELECT VersionNr FrOM  dbo.TestDbVersion";
+                    int version;
+                    using (SqlConnection connection = new SqlConnection(constr))
                     {
                         connection.Open();
+                        var cmd = connection.CreateCommand();
+                        cmd.CommandType = System.Data.CommandType.Text;
+                        cmd.CommandText = testCommand;
+                        using (var rdr = cmd.ExecuteReader())
+                        {
+                            rdr.Read();
+                            version = rdr.GetInt32(0);
+                        }
+                    }
 
-                        SqlCommand command = new SqlCommand("DROP DATABASE GenericBackendTest", connection);
-                        command.ExecuteNonQuery();
-                    }*/
+                    if (version < 2)
+                    {
+                        using (SqlConnection connection = new SqlConnection(@"server=(localdb)\MSSQLLocalDB"))
+                        {
+                            connection.Open();
+                            var cmdDropCon = new SqlCommand("ALTER DATABASE [GenericBackendTest] SET SINGLE_USER WITH ROLLBACK IMMEDIATE", connection);
+                            cmdDropCon.ExecuteNonQuery();
+                            SqlCommand command = new SqlCommand("DROP DATABASE GenericBackendTest", connection);
+                            command.ExecuteNonQuery();
+                        }
+                    }
+                    else
+                    {
+                        return; // Everything ok
+                    }
                 }
